+++ owrx/config/defaults.py
@@ -483,5 +483,8 @@ defaultConfig = PropertyLayer(
     lorawan_bw="7",
     meshtastic_bw="8",
     meshcore_bw="6",
-    meshcom_bw="8"
+    meshcom_bw="8",
+    speech_url="",
+    speech_squelch=20,
+    speech_hang_time=5000
 ).readonly()